feat: expand test coverage, add trajectory system, improve embedded support
Test infrastructure:
- Add 6 new test files: test_board.c, test_rng.c, test_pattern.c,
  test_cvd.c, test_effect.c, test_trajectory.c
- Total test coverage: 306 tests across 15 test files

Core improvements:
- Add trajectory computation system (pb_trajectory.h/c)
- Add pb_freestanding.h for embedded environments (pure integer math)
- Implement pb_board_insert_row for survival mode pressure
- Add expected_pops/drops calculation to solver
- Make bubble_radius configurable via ruleset

Cross-compilation:
- Fix mcs51/mos6502 function pointer issues (MICRO_SRCS)
- Add pure integer trig for freestanding targets (pb_fp_sin/cos/sqrt)
- Size tier support for memory-constrained platforms

Documentation:
- Add level-format-spec.md, special-bubbles-spec.md, trajectory-spec.md
- Update README with current test count

+++ pb_core/src/core/pb_replay.c
@@ -7,7 +7,7 @@
 
 #include "pb/pb_replay.h"
 #include <stdlib.h>
-#include <string.h>
+
 #include <stdio.h>
 
 /*============================================================================
@@ -51,10 +51,11 @@ int pb_varint_decode(const uint8_t* data, int len, uint32_t* out_value)
 
 int pb_varint_size(uint32_t value)
 {
-    if (value < (1 << 7)) return 1;
-    if (value < (1 << 14)) return 2;
-    if (value < (1 << 21)) return 3;
-    if (value < (1 << 28)) return 4;
+    /* Use uint32_t casts to avoid shift overflow on 8/16-bit platforms */
+    if (value < ((uint32_t)1 << 7)) return 1;
+    if (value < ((uint32_t)1 << 14)) return 2;
+    if (value < ((uint32_t)1 << 21)) return 3;
+    if (value < ((uint32_t)1 << 28)) return 4;
     return 5;
 }
